requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.2
pydantic-settings==2.8.1
python-dotenv==1.0.1
httpx==0.27.2
PyYAML==6.0.2
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
